Added todo's and removed TreeViewResults as it's no longer needed
