+++ IDesign/IDesign.Extension/UserControls/ExpanderResults.xaml.cs
@@ -20,5 +20,53 @@ namespace IDesign.Extension.UserControls
         {
             InitializeComponent();
         }
+
+        //TODO: Handle clicking on a node to bring the user to the right document
+        private void EventSetter_OnHandler(object sender, MouseButtonEventArgs e)
+        {
+            var viewItem = sender as TreeViewItem;
+
+            if (!(viewItem?.DataContext is CheckResultViewModel viewModel))
+            {
+                return;
+            }
+
+            var element = viewModel.Result.GetElement();
+            if (element == null)
+            {
+                return;
+            }
+
+            var node = element.GetSyntaxNode();
+            if (node == null)
+            {
+                return;
+            }
+
+            SelectNodeInEditor(node, element.GetRoot().GetSource());
+        }
+
+        /// <summary>
+        ///     Clicking on the node brings you to the right document.
+        /// </summary>
+        private void SelectNodeInEditor(SyntaxNode node, string file)
+        {
+            try
+            {
+                var tm = (IVsTextManager)Package.GetGlobalService(typeof(SVsTextManager));
+                var cm = (IComponentModel)Package.GetGlobalService(typeof(SComponentModel));
+                var ws = (Workspace)cm.GetService<VisualStudioWorkspace>();
+                var did = ws.CurrentSolution.GetDocumentIdsWithFilePath(file);
+                ws.OpenDocument(did.FirstOrDefault());
+                tm.GetActiveView(1, null, out var av);
+                var sp = node.GetLocation().GetMappedLineSpan().StartLinePosition;
+                var ep = node.GetLocation().GetMappedLineSpan().EndLinePosition;
+                av.SetSelection(sp.Line, sp.Character, ep.Line, ep.Character);
+            }
+            catch (Exception e)
+            {
+                _ = e.Message;
+            }
+        }
     }
 }
